art: Enable actual test
The caching needed a lot of work before it was ready.
Caches must not be owned by anyone other than the default user, as it
archives with owner info _and_ tries to restore them as was _without_
being root. ALso there were possibly relevant bugs filed against the
newest cache action, so pin a slightly older version.

The checkout action doesn't play nice with our chroot setup, so we need
to do this ourselves.