art: Finish jobs independently
So that a single failing job/arch doesn't block getting results for
all other arches. This is especially important for ports and the sid
suite, as they may have temporarily broken packages, resolving itself
after a day or so. Unfortunately GHA currently don't have a way to mark
such cases as "neutral" or similar, so we can only let them fail
graceless.