art: don't use dist-upgrade
This is the recommended way for stable releases anyway.
On Sid this may block some package transitions, but since the chroots
are regenerated every 45 days anyway, this shouldn't be too much of a
problem.
This should resolve the current RISC-V failures due to conflicting git
and git-man versions in its port archive.