art: do not recreate all chroots simultaneously
It often lead to the network connection of one or two jobs timing out
and causing failures. Let's hope this avoids the problem.

Also, this now makes use of the shell's arithmetic expansion.
Previously I mistakenly believed this to not be part of POSIX, but
it actually is since at least issue 6 (which MacOS is certified for).