art: cache VM images
This won't speed up the jobs, but it will make them
more reliable as the initial download is error-prone.

To achieve a speed up, the state after upgrade and package
installation would need to be cached. However, since vagrant doesn't
enable VirtualBox's discard on trim feature, those images would
together be too large to cache, with a significant portion of the image
(especially the OpenIndiana one) being used for no longer referenced
sectors.