art/vm: add non-kvm qemu runner
As a replacement for the now unreliable MacOS runners.
This does not user KVM, but in return it is hopefully more stable.
And performance isn’t thaaat bad when also taking into account the
time wasted before with and due to vagrant tasks and limitations.

Unfortunately, cross-platform-actions/action does not yet support
OpenIndiana or any other illumos (or Solaris) system.